using the user's input to qery the database.
I used the user's input through the name properties in the submit
event and the $_POST super global to create a SQL query to the
database. The query filers for the tag field using the syntax
SELET * FROM posts LIKE %$variable%
and prining the assoc array returned by the database.